# api/openapi/bondexchange/v1/bond_exchange.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Bond Exchange API",
    "description": "REST and gRPC API for publishing and buying sale offers.",
    "version": "3.0.0"
  },
  "tags": [
    {
      "name": "BondExchangeService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/active-bond-series": {
      "get": {
        "summary": "ListActiveBondSeries returns every bond series having at least one active offer.",
        "operationId": "BondExchangeService_ListActiveBondSeries",
        "responses": {
          "200": {
            "description": "The bond series having active offers.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.ListActiveBondSeriesResponse"
            }
          },
          "429": {
            "description": "The authenticated principal exhausted its shared 100-request allowance for the current UTC minute. Retry-After gives whole seconds until reset.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            },
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the next database-clock minute.",
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "500": {
            "description": "The server could not complete the request.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          }
        },
        "tags": [
          "BondExchangeService"
        ]
      }
    },
    "/active-offers": {
      "get": {
        "summary": "ListActiveOffers streams every active sale offer from one database snapshot.",
        "operationId": "BondExchangeService_ListActiveOffers",
        "responses": {
          "200": {
            "description": "The active sale offers.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.ListActiveOffersResponse"
            }
          },
          "400": {
            "description": "The required bond series is missing or invalid.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          },
          "429": {
            "description": "The authenticated principal exhausted its shared 100-request allowance for the current UTC minute. Retry-After gives whole seconds until reset.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            },
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the next database-clock minute.",
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "500": {
            "description": "The server could not complete the request.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          }
        },
        "parameters": [
          {
            "name": "bond",
            "description": "Required bond series to list. The service canonicalizes this value to uppercase.",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BondExchangeService"
        ]
      }
    },
    "/buys": {
      "post": {
        "summary": "Buy places a binding order or reservation for one currently active sale offer.\nSettlement, payment, custody, and ownership transfer are outside this API.",
        "operationId": "BondExchangeService_Buy",
        "responses": {
          "201": {
            "description": "The offer was purchased.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.BuyResponse"
            }
          },
          "400": {
            "description": "The request is invalid.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          },
          "404": {
            "description": "The active offer is unavailable.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          },
          "429": {
            "description": "The authenticated principal exhausted its shared 100-request allowance for the current UTC minute. Retry-After gives whole seconds until reset.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            },
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the next database-clock minute.",
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "500": {
            "description": "The server could not complete the request.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.BuyRequest"
            }
          },
          {
            "name": "Idempotency-Key",
            "description": "A canonical UUIDv4 nonce bound into the operation assertion.",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BondExchangeService"
        ]
      }
    },
    "/event-publications:publish-pending": {
      "post": {
        "summary": "PublishPendingEvents explicitly retries pending integration-event deliveries.\nNo automatic startup or scheduled drain invokes this operation.",
        "operationId": "BondExchangeService_PublishPendingEvents",
        "responses": {
          "200": {
            "description": "The visible pending events were attempted.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.PublishPendingEventsResponse"
            }
          },
          "400": {
            "description": "The destination is invalid or no publisher is configured.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          },
          "429": {
            "description": "The authenticated principal exhausted its shared 100-request allowance for the current UTC minute. Retry-After gives whole seconds until reset.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            },
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the next database-clock minute.",
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "500": {
            "description": "The server could not complete the recovery attempt.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.PublishPendingEventsRequest"
            }
          },
          {
            "name": "Idempotency-Key",
            "description": "A canonical UUIDv4 nonce bound into the operation assertion; per-event leases coordinate overlapping drains.",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BondExchangeService"
        ]
      }
    },
    "/healthz": {
      "get": {
        "summary": "CheckHealth reports whether the server can reach its database.",
        "operationId": "BondExchangeService_CheckHealth",
        "responses": {
          "200": {
            "description": "The service is healthy.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.CheckHealthResponse"
            }
          },
          "429": {
            "description": "The authenticated principal exhausted its shared 100-request allowance for the current UTC minute. Retry-After gives whole seconds until reset.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            },
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the next database-clock minute.",
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "503": {
            "description": "The database is unavailable.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          }
        },
        "tags": [
          "BondExchangeService"
        ]
      }
    },
    "/sale-offer-quotes": {
      "post": {
        "summary": "QuoteSaleOffer creates an expiring USD-to-MXN FIX quote for explicit seller acceptance.",
        "operationId": "BondExchangeService_QuoteSaleOffer",
        "responses": {
          "201": {
            "description": "The expiring conversion quote was created.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.QuoteSaleOfferResponse"
            }
          },
          "400": {
            "description": "The quote request is invalid.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          },
          "404": {
            "description": "The bond series does not exist.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          },
          "409": {
            "description": "The idempotency nonce was reused for a different request.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          },
          "429": {
            "description": "The authenticated principal exhausted its shared 100-request allowance for the current UTC minute. Retry-After gives whole seconds until reset.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            },
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the next database-clock minute.",
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "503": {
            "description": "An acceptable Banxico FIX observation is unavailable.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.QuoteSaleOfferRequest"
            }
          },
          {
            "name": "Idempotency-Key",
            "description": "A canonical UUIDv4 nonce bound into the operation assertion.",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BondExchangeService"
        ]
      }
    },
    "/sale-offers": {
      "post": {
        "summary": "CreateSaleOffer publishes immutable canonical MXN terms for the authenticated principal.",
        "operationId": "BondExchangeService_CreateSaleOffer",
        "responses": {
          "201": {
            "description": "The sale offer was created.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.CreateSaleOfferResponse"
            }
          },
          "400": {
            "description": "The sale offer is invalid or its required conversion quote cannot be accepted.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          },
          "404": {
            "description": "The bond series does not exist.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          },
          "409": {
            "description": "The generated sale-offer ID collided with an existing resource.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          },
          "429": {
            "description": "The authenticated principal exhausted its shared 100-request allowance for the current UTC minute. Retry-After gives whole seconds until reset.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            },
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the next database-clock minute.",
                "type": "integer",
                "format": "int64"
              }
            }
          },
          "500": {
            "description": "The server could not complete the request.",
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.Error"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/bondexchange.v1.CreateSaleOfferRequest"
            }
          },
          {
            "name": "Idempotency-Key",
            "description": "A canonical UUIDv4 nonce bound into the operation assertion.",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BondExchangeService"
        ]
      }
    }
  },
  "definitions": {
    "bondexchange.v1.BuyRequest": {
      "type": "object",
      "properties": {
        "sale_offer_id": {
          "type": "string",
          "description": "Canonical UUIDv7 returned by the service when the offer was created."
        }
      },
      "required": [
        "sale_offer_id"
      ]
    },
    "bondexchange.v1.BuyResponse": {
      "type": "object",
      "properties": {
        "offer": {
          "$ref": "#/definitions/bondexchange.v1.SaleOffer"
        },
        "bought_at": {
          "type": "string",
          "format": "date-time"
        }
      },
      "required": [
        "offer",
        "bought_at"
      ]
    },
    "bondexchange.v1.CheckHealthResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        }
      },
      "required": [
        "status"
      ]
    },
    "bondexchange.v1.CreateSaleOfferRequest": {
      "type": "object",
      "properties": {
        "bond_series": {
          "type": "string",
          "description": "Bond series to sell. The service canonicalizes this value to uppercase."
        },
        "price": {
          "type": "string",
          "description": "Exact positive submitted amount encoded as a string."
        },
        "currency_code": {
          "type": "string",
          "description": "MXN creates directly. USD requires an accepted conversion_quote_id."
        },
        "conversion_quote_id": {
          "type": "string",
          "description": "Required for USD submissions and forbidden for MXN submissions."
        }
      },
      "required": [
        "bond_series",
        "price",
        "currency_code"
      ]
    },
    "bondexchange.v1.CreateSaleOfferResponse": {
      "type": "object",
      "properties": {
        "offer": {
          "$ref": "#/definitions/bondexchange.v1.SaleOffer"
        }
      },
      "required": [
        "offer"
      ]
    },
    "bondexchange.v1.Error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      },
      "description": "Error is the REST error response. Native gRPC callers receive the equivalent status code and message.",
      "required": [
        "error"
      ]
    },
    "bondexchange.v1.ListActiveBondSeriesResponse": {
      "type": "object",
      "properties": {
        "bond_series": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "bond_series"
      ]
    },
    "bondexchange.v1.ListActiveOffersComplete": {
      "type": "object",
      "properties": {
        "offer_count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "bondexchange.v1.ListActiveOffersResponse": {
      "type": "object",
      "properties": {
        "offer": {
          "$ref": "#/definitions/bondexchange.v1.SaleOffer"
        },
        "complete": {
          "$ref": "#/definitions/bondexchange.v1.ListActiveOffersComplete"
        }
      }
    },
    "bondexchange.v1.PublishPendingEventsRequest": {
      "type": "object",
      "properties": {
        "destination_id": {
          "type": "string",
          "description": "Stable configured destination identifier. Empty selects every destination."
        }
      }
    },
    "bondexchange.v1.PublishPendingEventsResponse": {
      "type": "object",
      "properties": {
        "attempted": {
          "type": "string",
          "format": "uint64"
        },
        "delivered": {
          "type": "string",
          "format": "uint64"
        },
        "failed": {
          "type": "string",
          "format": "uint64"
        },
        "remaining": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "bondexchange.v1.QuoteSaleOfferRequest": {
      "type": "object",
      "properties": {
        "bond_series": {
          "type": "string"
        },
        "price": {
          "type": "string",
          "description": "Exact positive USD amount encoded as a string."
        },
        "currency_code": {
          "type": "string",
          "description": "Must be USD. MXN offers do not require a conversion quote."
        }
      },
      "required": [
        "bond_series",
        "price",
        "currency_code"
      ]
    },
    "bondexchange.v1.QuoteSaleOfferResponse": {
      "type": "object",
      "properties": {
        "quote_id": {
          "type": "string"
        },
        "bond_series": {
          "type": "string"
        },
        "submitted_price": {
          "type": "string"
        },
        "submitted_currency_code": {
          "type": "string"
        },
        "mxn_price": {
          "type": "string"
        },
        "currency_code": {
          "type": "string"
        },
        "rate": {
          "type": "string"
        },
        "rate_series": {
          "type": "string"
        },
        "rate_observed_on": {
          "type": "string",
          "format": "date-time"
        },
        "expires_at": {
          "type": "string",
          "format": "date-time"
        }
      },
      "required": [
        "quote_id",
        "bond_series",
        "submitted_price",
        "submitted_currency_code",
        "mxn_price",
        "currency_code",
        "rate",
        "rate_series",
        "rate_observed_on",
        "expires_at"
      ]
    },
    "bondexchange.v1.SaleOffer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "PostgreSQL-generated canonical UUIDv7."
        },
        "bond_series": {
          "type": "string"
        },
        "price": {
          "type": "string",
          "description": "Exact decimal monetary amount encoded as a string."
        },
        "currency_code": {
          "type": "string"
        }
      },
      "required": [
        "id",
        "bond_series",
        "price",
        "currency_code"
      ]
    }
  },
  "securityDefinitions": {
    "OperationAssertion": {
      "type": "apiKey",
      "description": "A Bearer JWT bound to the operation and deterministic protobuf request digest.",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "OperationAssertion": []
    }
  ]
}